+++ packages/devkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @slicemedia/devkit
 
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @slicemedia/devkit-core@0.3.0
+  - @slicemedia/devkit-addon@0.3.0
+
 ## 0.2.0
 
 ### Patch Changes